docshell/test/navigation/test_bug1745638.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>bug 1745638</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
  // This test triggers an assertion in the old session history
  // implementation.
  SimpleTest.expectAssertions(0, 1);

  SimpleTest.waitForExplicitFinish();
  var testWindow;
  var loadCount = 0;
  function test() {
    testWindow = window.open("file_bug1745638.html");
  }

  function pageLoaded() {
    ++loadCount;
    is(testWindow.document.getElementById('testFrame').contentDocument.body.innerHTML,
       "passed",
       "Iframe's textual content should be 'passed'.");
    if (loadCount == 1) {
      testWindow.history.go(0);
    } else {
      testWindow.close();
      SimpleTest.finish();
    }
  }

  </script>
</head>
<body onload="test()">
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
